comments, set day schedule function, repeat function
added setting one day's schedule
added some documentary
repeat function for set_schedule

+++ schedule/models.py
@@ -10,6 +10,14 @@ from django.core.mail import send_mail
 
 # Create your models here.
 class Shift(models.Model):
+    """The shift model is suppose to register common shifts. Ie morning/night 
+    shifts, to avoid keep typing the same thing. It becomes a reusable object
+    
+    Shift start is starting time for the shift, and duration is how long
+    the shift will last for
+    
+    Shift name is just a reference name ie morning shift
+    """
     # correspond to datetime.time object
     shift_start = models.TimeField(help_text='enter format hh:mm:ss') 
     # correspond to datetime.timedelta
@@ -20,10 +28,23 @@ class Shift(models.Model):
         return self.shift_name
     
 class Schedule(models.Model):
-        
+    """
+    The schedule model is suppose to register common schedule pattern. For
+    example, if morning/morning/night/night/rest/rest/rest/rest days are common
+    pattern, then it can be stored as a pattern, and admin does not have to repeat
+    this schedule pattern
+    
+    cycle is how many days does this cycle cut off? In the aforementioned 
+    shift cycle, it is 8 days. There will be a lot of day fields to accommodate
+    different shift patterns. But if there are 10 day fields, and cut off is 
+    at 8, then it will be a cycle of the first 8 day. The last two are ignored
+    
+    
+    """
     schedule_name = models.CharField(max_length=20, primary_key=True)
     cycle = models.PositiveSmallIntegerField(MaxValueValidator(10),
                                              help_text='how many days is the shift pattern cycle')
+
     day_1 = models.ForeignKey(Shift, 
                               related_name='%(app_label)s_%(class)s_1',
                              on_delete=models.SET_NULL, 
@@ -49,6 +70,16 @@ class Schedule(models.Model):
     
 
 class Assign(models.Model):
+    """Assign model is made to assign schedule to individual. It will store
+    all schedule in this model. You can filter for individual, day, etc
+    
+    field switch is to indicate whether this schedule is currently hoping 
+    for a switch. Once a swap is made, this will be made false
+    
+    
+    Will need to have a mass schedule assign function, where it can apply 
+    pattern to all individual at once
+    """
     # correspond to datetime.date
     start_date = models.DateField(help_text='start day for shift')
     
@@ -71,7 +102,16 @@ class Assign(models.Model):
         return result
         
 class Request(models.Model):
-    requester = models.ForeignKey(Individual,
+    """This Request model is designed to hold those schedules that are currently
+    requesting the receiver for a swap. It happens after the requester selects
+    who to make the switch with.
+    
+    Accept is whether receiver accepts the shift swap
+    done is whether a response has been accepted
+    
+    
+    """
+    applicant = models.ForeignKey(Individual,
                                   on_delete=models.CASCADE,
                                   null=True,
                                   related_name="requester")
@@ -79,45 +119,106 @@ class Request(models.Model):
                                  on_delete=models.CASCADE,
                                  null=True,
                                  related_name="receiver")
-    requester_shift = models.ForeignKey(Assign,
+    applicant_schedule = models.ForeignKey(Assign,
                               on_delete=models.SET_NULL,
                               null=True,
                               related_name="requester_shift")
-    receiver_shift = models.ForeignKey(Assign,
+    receiver_schedule = models.ForeignKey(Assign,
                                        on_delete=models.SET_NULL,
                                        null=True,
                                        related_name="receiver_shift")
+    # get creation timestamp for timeout purpose later
+    created = models.DateTimeField(auto_now_add=True)
+    
     # did the receiver accept the proposed swap?
     accept = models.BooleanField(default=False)
     # is the action done?
     done = models.BooleanField(default=False)
-#    
+
     
 def get_schedule(person):
-    """get a person's schedules sorted by date"""
+    """
+    get a person's schedules sorted by date
+    
+    person is Individual instance"""
     existing_schedule = Assign.objects.filter(individual__exact=person).order_by('shift_start')
 #    print('fetching schedule for', person.user.first_name, 
 #          person.user.last_name)
 #    for schedule in existing_schedule:
 #        print(schedule.shift_start, schedule.shift_end)
     return existing_schedule
-        
-        
-def set_schedule(person, start_date, shift_pattern):
+
+def set_schedule_day(person, start_day, shift):
     """
+    set schedule for a single day on that person
+    
+    person is Individual instance
+    start_day is datetime.date
+    shift is Shfit instance"""
+
+    shift_time = pytz.UTC.localize(shift.shift_start)
+    shift_dur = shift.shift_duration
+
+    # shift_start and shift_end tzinfo is UTC
+    shift_start = datetime.datetime.combine(start_day, shift_time)
+    shift_end = shift_start + shift_dur
+        
+    # register dates
+    # check if there are any shift conflicts
+    existing_schedule = Assign.objects.filter(individual__exact=person).filter(start_date__exact=start_day)
+  
+    if not existing_schedule:
+        schedule = Assign(start_date = start_day,
+                          shift_start = shift_start,
+                          shift_end = shift_end,
+                          individual = person)
+        schedule.save()
+        status = True
+    else:
+        # else if there is already a schedule
+        # find all schedule time
+        overlap = False
+        for exist_shift in existing_schedule:
+            overlap = exist_shift.shift_start <= shift_start < exist_shift.shift_end
+            if overlap:
+                # if found an overlap
+                break
+            
+        if overlap:
+            # if there is overlap of shifts
+            print('registraion failed due to schedule conflict')
+            status = False
+        else:
+            # if no overlap (but allow double shifts ie shift continuation)
+            schedule = Assign(start_date = start_day,
+                          shift_start = shift_start,
+                          shift_end = shift_end,
+                          individual = person)
+            schedule.save()
+            status = True
+    if status:
+        print('schedule resgistered')
+    return status
+    
+    
+def set_schedule(person, start_date, shift_pattern, repeat=1):
+    """
+    This function sets a person's schedule
+    
+    person is Individual instance
     start_date is datetime.date object
-    shift is shift pattern 
+    shift is shift pattern from Schedule
     """
     
     # shift_start is datetime.time
     not_registered = []
-    working_days = len(shift_pattern.day_list)
+    working_days = len(shift_pattern.day_list)*repeat
     working_dates = []
     for i in range(working_days):
         working_dates.append(start_date + datetime.timedelta(days=i))
 
 
-    for pattern, dates in zip(shift_pattern.day_list, working_dates):
+    for pattern, dates in zip(shift_pattern.day_list*repeat, working_dates):
         if pattern is None:
             # if it is a rest day, go to next iteration to set schedule
             continue
@@ -176,19 +277,35 @@ def clear_schedule(person):
     
     while True:
         try:
-            option = input("This will delete all the person's schedule. \
-                           Choose an option to proceed [y/n]: ").lower()
+            option = input("This will delete all the person's schedule.\
+                           Choose an option to proceed [y/N]: ").lower()
             if option in ['y','n']:
                 break
         except:
             pass
     if option == 'y':
-        Assign.objects.filter(individual__exact=person).delete()
+        while True:
+            try:
+                employee_id = input("The schedule will be permanently deleted. \
+                                    Are you sure? Enter employee ID if you\
+                                    want to delete schedule: ").lower()
+                
+                if employee_id.isdigit() and int(employee_id) == person.employee_id:
+                    Assign.objects.filter(individual__exact=person).delete()
+                    break
+                else:
+                    print('employee ID does not match, will not delete any schedule')
+                    break
+            except:
+                pass
     else:
         print("person's schedule not deleted")
         
 def del_schedule(person, date):
-    """delete a person's schedule on a particular date"""
+    """delete a person's schedule on a particular date
+    person is Individual instance
+    date is datetime.date
+    """
     schedule = Assign.objects.filter(individual__exact=person).filter(start_date__exact=date)
     if schedule.count() > 1:
         # if there are more than one schedule on that date
@@ -212,10 +329,34 @@ def del_schedule(person, date):
     
         
 def swap(person, swap_shift_start):
-    """first checks people that are currently swapping,
+    """
+    this function will check if swap shift is possible
+    person is Individual instance, and the requester
+    swap_shift_start is the shift that needs to be swapped
+    
+    This will output a dictionary with key success, available_shifts, free_people
+    
+    first checks people that are currently swapping,
     then check people that are open to swap (even if not currently swapping)
     then check people that are open to swap but are not working (no shifts offered
-    in return. Acceptor does not offer shift in return on last case"""
+    in return. Acceptor does not offer shift in return on last case
+    
+    if a shift/person is found to be possible to make swap, success will 
+    be True. If there are shifts offered in return, the available_shifts 
+    will be QuerySet. If there are no shifts offered in return, then, 
+    available_shifts will remain as None and free_people will return as a list 
+    of Individual instance who are open to accept shift and are not working 
+    that shift
+    
+    If not possible to find any shift/person to swap, success will be Fail
+    and available_shifts will be None and free_people will be empty list
+    
+    The current logic of swap shift does not allow double shifts (inclusive). 
+    If person A wants to swap 7-9 shift and person B can offer 9-11. It will 
+    not allow the shift to be swapped
+    
+    
+    """
     print(person)
     success = True
     output = None
@@ -310,7 +451,10 @@ def send_email(subject,
                msg,
                sender_address,
                receiver_list):
-    """subject, msg, sender_address in str,
+    """
+    This function will send email
+    
+    subject, msg, sender_address in str,
     receiver_list in list of str"""
     success = send_mail(subject=subject,
               message=msg,
